% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_procedures}
\alias{acs_procedures}
\title{Contains angiograms,pci, and cabg for acs cohort}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 394 rows and 16 columns.
}
\usage{
acs_procedures
}
\description{
Contains angiograms,pci, and cabg for acs cohort
}
\keyword{datasets}
