% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{business_hours}
\alias{business_hours}
\title{Definition for an admission during business hours}
\usage{
business_hours()
}
\value{
a vector of strings
}
\description{
Definition for an admission during business hours
}
\seealso{
Other definitions: 
\code{\link{ACS_codes}()},
\code{\link{CCLHD_facility}()},
\code{\link{NSLHD_facility}()},
\code{\link{exclusion_criteria_meds}()},
\code{\link{hf_icd_codes}()},
\code{\link{internal_transfers}()},
\code{\link{separation_modes}()}
}
\concept{definitions}
