% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{CCLHD_facility}
\alias{CCLHD_facility}
\title{Returns the definition of CCLHD facilities}
\usage{
CCLHD_facility()
}
\value{
a vector of strings
}
\description{
Returns the definition of CCLHD facilities
}
