% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_demographics}
\alias{create_demographics}
\title{Creates demographic information}
\usage{
create_demographics(journeys, form_triage_classified)
}
\arguments{
\item{journeys}{a df containing journeys}

\item{form_triage_classified}{df containing triage forms}
}
\value{
a df containing journeys and the additional columns
}
\description{
Creates columns such as \code{AGE}, \code{ADMIT_IN_BUSINESS_HOURS}, \code{LENGTH_OF_STAY}
for each journey
}
\seealso{
Other journey: 
\code{\link{create_discharge_letter_keys}()},
\code{\link{create_interventions}()},
\code{\link{create_meds_admit_discharge}()},
\code{\link{create_presentations}()},
\code{\link{create_separations}()},
\code{\link{create_trajectory}()},
\code{\link{extract_triage_forms}()}
}
\concept{journey}
