% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_discharge_letter_keys}
\alias{create_discharge_letter_keys}
\title{Extracts cohort discharge keys}
\usage{
create_discharge_letter_keys(journeys, discharge_letter_keys)
}
\description{
Extracts cohort discharge keys
}
\seealso{
Other journey: 
\code{\link{create_demographics}()},
\code{\link{create_interventions}()},
\code{\link{create_meds_admit_discharge}()},
\code{\link{create_presentations}()},
\code{\link{create_separations}()},
\code{\link{create_trajectory}()},
\code{\link{extract_triage_forms}()}
}
\concept{journey}
