% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_door_to_balloon}
\alias{create_door_to_balloon}
\title{Extracts door to balloon times}
\usage{
create_door_to_balloon(journeys, door_to_balloon, demographics)
}
\arguments{
\item{journeys}{a df containing journeys}

\item{door_to_balloon}{a df containing balloon times}

\item{demographics}{a df containing demopgrahic information
of journeys such as \code{TRAJECTORY} and \code{PRESENTATION_MODE}}
}
\value{
the door_to_balloon df and additional columns \code{DTB_PRESENTATION_MODE}
and \code{DELTA_BINNED} for the cohort journeys
}
\description{
Extracts door to balloon times,  bins the times
and assigns a category to each door to baloon measurement
}
\seealso{
Other acs: 
\code{\link{create_meds_discharge}()},
\code{\link{create_stemi_code}()},
\code{\link{extract_acs_procedures}()},
\code{\link{merge_mckesson_cerner_procedures}()},
\code{\link{select_acs_cohort}()}
}
\concept{acs}
