% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartfailure.R
\name{create_hf_meds_discharge}
\alias{create_hf_meds_discharge}
\title{Creates discharge medication}
\usage{
create_hf_meds_discharge(journeys, meds, discharge_letter_keys)
}
\arguments{
\item{journeys}{a df containing journeys}

\item{meds}{a df containing medication}

\item{discharge_letter_keys}{a df containing keys of encounters with discharge letters}
}
\value{
a df summarising the counts of the specified medication for each journey
}
\description{
Summarises discharge medications for ACS cohort. In particular, summarises
\code{Diuretic},\code{A2RB},\verb{Beta blockers} and \verb{Beta blocker}
for each journey.
}
\details{
It excludes journeys that meet the exclusion criteria, see \code{\link{exclusion_criteria_meds}}
and includes journeys with a discharge letter key.
}
\seealso{
\code{\link{create_meds_discharge}} for acs medication

Other heart failure: 
\code{\link{create_hf_icd10_encounters}()},
\code{\link{create_hf_snomed_encounters}()},
\code{\link{extract_echo_procedures}()},
\code{\link{select_hf_cohort}()}
}
\concept{heart failure}
