% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartfailure.R
\name{create_hf_snomed_encounters}
\alias{create_hf_snomed_encounters}
\title{Extracts SNOMED heart failure encounters}
\usage{
create_hf_snomed_encounters(
  encounters,
  diagnosis,
  journey_start_begin = "2017-04-01",
  journey_start_end = "2017-06-30"
)
}
\arguments{
\item{diagnosis}{a df containing diagnosis}

\item{journey_start_begin}{start date for the cohort}

\item{journey_start_end}{end date}

\item{journey}{a df containing journeys}
}
\description{
Extracts the encounters with SNOMED heart failure diagnosis
with \code{JOURNEY_START} between \code{journey_start_begin} and \code{journey_start_end}
}
\seealso{
\code{\link{select_hf_cohort}} for selection of journey cohort
based on ICD10 codes.

Other heart failure: 
\code{\link{create_hf_icd10_encounters}()},
\code{\link{create_hf_meds_discharge}()},
\code{\link{extract_echo_procedures}()},
\code{\link{select_hf_cohort}()}
}
\concept{heart failure}
