% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_interventions}
\alias{create_interventions}
\title{Creates categories for the types of interventions
Patients are classified based on whether they have had
any of the following procedures: angiogram,pci,CABG
or whether they were transferred to a private hospital}
\usage{
create_interventions(journeys, procedures)
}
\arguments{
\item{journeys}{a df containing encounters and their journey keys}

\item{procedures}{a df containing the relevant procedures for the cohort}
}
\value{
a df containing, \code{INTERVENTION}
}
\description{
It is assumed that if a patient had a PCI, or CABG
they would have an angiogram
}
