% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_meds_admit_discharge}
\alias{create_meds_admit_discharge}
\title{Extracts medications relevant to the cohort journeys}
\usage{
create_meds_admit_discharge(journey, journey_meds_admit_discharge)
}
\arguments{
\item{journey}{a df containin journeys}

\item{journey_meds_admit_discharge}{a df containing admission and discharge
medication}
}
\value{
a df containing the medications for the cohort journeys
}
\description{
Extracts medications relevant to the cohort journeys
}
\seealso{
Other journey: 
\code{\link{create_demographics}()},
\code{\link{create_discharge_letter_keys}()},
\code{\link{create_interventions}()},
\code{\link{create_presentations}()},
\code{\link{create_separations}()},
\code{\link{create_trajectory}()},
\code{\link{extract_triage_forms}()}
}
\concept{journey}
