% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_meds_discharge}
\alias{create_meds_discharge}
\title{Creates discharge medication}
\usage{
create_meds_discharge(journeys, meds, discharge_letter_keys)
}
\arguments{
\item{journeys}{a df containing journeys}

\item{meds}{a df containing medication}

\item{discharge_letter_keys}{a df containing keys of encounters with discharge letters}
}
\value{
a df summarising the counts of the specified medication for each journey
}
\description{
Summarises discharge medications for ACS cohort. In particular, summarises
\code{aspirin},\code{P2Y12},\code{Anticoagulant},\code{A2RB},\verb{Beta blockers} and \code{Statins}
for each journey.
}
\details{
It excludes journeys that meet the exclusion criteria, see \code{\link{exclusion_criteria_meds}}
and includes journeys with a discharge letter key.
}
\seealso{
\code{\link{create_hf_meds_discharge}} for heart failure medication

Other acs: 
\code{\link{create_door_to_balloon}()},
\code{\link{create_stemi_code}()},
\code{\link{extract_acs_procedures}()},
\code{\link{merge_mckesson_cerner_procedures}()},
\code{\link{select_acs_cohort}()}
}
\concept{acs}
