% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_separations}
\alias{create_separations}
\title{Defines the separation mode for journeys and groups separation modes
that are not of interest to "Other"}
\usage{
create_separations(journeys)
}
\arguments{
\item{journeys}{a df containing encounters and their journey keys}
}
\value{
a df containing, \code{SEPARATION_MODE}
}
\description{
Defines the separation mode for journeys and groups separation modes
that are not of interest to "Other"
}
