% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_stemi_code}
\alias{create_stemi_code}
\title{Creates SNOMED code column}
\usage{
create_stemi_code(journeys, diagnosis)
}
\arguments{
\item{journeys}{a df containing journeys}

\item{diagnosis}{a df containing diagnosis codes}
}
\value{
the \code{journeys} df with additional logical columns \code{STEMI_SNOMED} and \code{NSTEMI_SNOMED}
describing whether that journeys had the relevant SNOMED code
}
\description{
Extracts any STEMI or NSTEMI  related SNOMED diagnosis for the cohort journey
}
\seealso{
Other acs: 
\code{\link{create_door_to_balloon}()},
\code{\link{create_meds_discharge}()},
\code{\link{extract_acs_procedures}()},
\code{\link{merge_mckesson_cerner_procedures}()},
\code{\link{select_acs_cohort}()}
}
\concept{acs}
