% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_trajectory}
\alias{create_trajectory}
\title{Categorises the trajectories of journeys}
\usage{
create_trajectory(journeys)
}
\arguments{
\item{journeys}{a df containing encounters and their journey keys}
}
\value{
journeys dataframe with the columns, \code{TRANSFER_STATUS} and \code{TRAJECTORY}
}
\description{
Categorises the trajectories of journeys and whether the journey was a transfer or not
Include a separate column for the presence of a coronary artery bypass procedure
}
\seealso{
Other journey: 
\code{\link{create_demographics}()},
\code{\link{create_discharge_letter_keys}()},
\code{\link{create_interventions}()},
\code{\link{create_meds_admit_discharge}()},
\code{\link{create_presentations}()},
\code{\link{create_separations}()},
\code{\link{extract_triage_forms}()}
}
\concept{journey}
