% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diagnosis_data}
\alias{diagnosis_data}
\title{Diagnosis for the encounters given by \code{encounter_data}}
\format{
An object of class \code{data.frame} with 721899 rows and 26 columns.
}
\usage{
diagnosis_data
}
\description{
A dataset containing diagnosises
}
\keyword{datasets}
