% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encounter_data}
\alias{encounter_data}
\title{Encounters for the years between 2016-2018 at RNS.}
\format{
An object of class \code{data.frame} with 602594 rows and 35 columns.
}
\usage{
encounter_data
}
\description{
A dataset containing encounters.
}
\keyword{datasets}
