% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{exclusion_criteria_meds}
\alias{exclusion_criteria_meds}
\title{Definition the exclusion criteria for medication}
\usage{
exclusion_criteria_meds()
}
\description{
Definition the exclusion criteria for medication
}
