% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartfailure.R
\name{extract_echo_procedures}
\alias{extract_echo_procedures}
\title{Extracts echocardiogram procedures}
\usage{
extract_echo_procedures(journeys, procedures)
}
\arguments{
\item{procedures}{a df containing procedures}

\item{journey}{a df containing journeys}
}
\description{
Extracts echo procedures based on a string search for the specified cohort
}
\seealso{
Other heart failure: 
\code{\link{create_hf_icd10_encounters}()},
\code{\link{create_hf_meds_discharge}()},
\code{\link{create_hf_snomed_encounters}()},
\code{\link{select_hf_cohort}()}
}
\concept{heart failure}
