% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{extract_triage_forms}
\alias{extract_triage_forms}
\title{Extracts ATAMI triage forms
An ATAMI triage form is any form that contains
\code{atami}, \code{etami},\verb{cath lab},\code{cathlab}}
\usage{
extract_triage_forms(form_triage_classified)
}
\arguments{
\item{form_triag_classified}{a df containing triag forms}
}
\value{
a dataframe that contains only ATAMI triag forms
}
\description{
Extracts ATAMI triage forms
An ATAMI triage form is any form that contains
\code{atami}, \code{etami},\verb{cath lab},\code{cathlab}
}
