% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{hf_icd_codes}
\alias{hf_icd_codes}
\title{Definition for heart failure ICD10 codes}
\usage{
hf_icd_codes()
}
\value{
a vector of strings
}
\description{
Definition for heart failure ICD10 codes
}
\seealso{
Other definitions: 
\code{\link{ACS_codes}()},
\code{\link{CCLHD_facility}()},
\code{\link{NSLHD_facility}()},
\code{\link{business_hours}()},
\code{\link{exclusion_criteria_meds}()},
\code{\link{internal_transfers}()},
\code{\link{separation_modes}()}
}
\concept{definitions}
