% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{internal_transfers}
\alias{internal_transfers}
\title{Definition of internal transfer}
\usage{
internal_transfers()
}
\value{
a vector of strings
}
\description{
Definition of internal transfer
}
\seealso{
Other definitions: 
\code{\link{ACS_codes}()},
\code{\link{CCLHD_facility}()},
\code{\link{NSLHD_facility}()},
\code{\link{business_hours}()},
\code{\link{exclusion_criteria_meds}()},
\code{\link{hf_icd_codes}()},
\code{\link{separation_modes}()}
}
\concept{definitions}
