% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journey_acs}
\alias{journey_acs}
\title{Contains journeys according to ACS rule 2}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1646 rows and 36 columns.
}
\usage{
journey_acs
}
\description{
\strong{Relative to absolute value:}
}
\details{
This rule is simpler than \strong{ACS Rule 1}, and simply involves comparing

If a result exceeds that, flag its row with \code{abs_high_thresh_flag <- TRUE},

and flag all rows in that journey with \code{group_pass_abs_high_thresh <- TRUE}
"journey_acs"
}
\keyword{datasets}
