% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journey_analysis_base}
\alias{journey_analysis_base}
\title{Contains journeys as defined by ACS for \code{encounter_data}}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 305633 rows and 36 columns.
}
\usage{
journey_analysis_base
}
\description{
Contains journeys as defined by ACS for \code{encounter_data}
}
\keyword{datasets}
