% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_table}
\alias{load_table}
\title{Loads the cache from SPEED-EXTRACT pipeline
and the specified table}
\usage{
load_table(
  table,
  config_dir = "D:/projects/SPEED-EXTRACT.pipeline/local.yaml",
  profile = "3m_hpc"
)
}
\arguments{
\item{table}{a string for the table to ne loaded}

\item{config_dir}{directory that contains the data profiles}

\item{profile}{strin that defines the data profile to be used}
}
\value{
a dataframe
}
\description{
Loads the cache from SPEED-EXTRACT pipeline
and the specified table
}
\concept{utils}
