% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{merge_mckesson_cerner_procedures}
\alias{merge_mckesson_cerner_procedures}
\title{Merges procedure table with McKesson data}
\usage{
merge_mckesson_cerner_procedures(procedures, door_to_balloon)
}
\arguments{
\item{procedures}{a df containing ICD10 procedures.}

\item{door_to_balloon}{as df containing doot to balloon times for McKesson
PCI procedures}
}
\value{
a df
}
\description{
Merges procedure table with McKesson data particularly for PCI
procedures, ensuring consistency between visualation of
ICD10 procedures and McKesson
}
\seealso{
Other acs: 
\code{\link{create_door_to_balloon}()},
\code{\link{create_meds_discharge}()},
\code{\link{create_stemi_code}()},
\code{\link{extract_acs_procedures}()},
\code{\link{select_acs_cohort}()}
}
\concept{acs}
