% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{NSLHD_facility}
\alias{NSLHD_facility}
\title{Returns the definition of NSLHD facilities}
\usage{
NSLHD_facility()
}
\value{
a vector of strings
}
\description{
Returns the definition of NSLHD facilities
}
