% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{select_acs_cohort}
\alias{select_acs_cohort}
\title{Extracts the ACS cohort}
\usage{
select_acs_cohort(
  journey_analysis_base,
  journey_start_begin = "2017-04-01",
  journey_start_end = "2017-06-30"
)
}
\arguments{
\item{journey_analysis_base, }{a dataframe containing a list of encounters with their journey_keys}

\item{journey_start_begin}{a date time string}

\item{journey_start_end}{a date time string}
}
\value{
a dataframe containing the ACS journeys
}
\description{
The cohort is defined as patients with a journey start date
between \code{journey_start_begin} and \code{journey_start_end}
and those with ICD10 STEMI/NSTEMI code.
It includes those who were admitted into ED OR patients who had transferred to RNSH
}
\seealso{
Other acs: 
\code{\link{create_door_to_balloon}()},
\code{\link{create_meds_discharge}()},
\code{\link{create_stemi_code}()},
\code{\link{extract_acs_procedures}()},
\code{\link{merge_mckesson_cerner_procedures}()}
}
\concept{acs}
