% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartfailure.R
\name{select_hf_cohort}
\alias{select_hf_cohort}
\title{Extracts heart failure cohort}
\usage{
select_hf_cohort(
  journey,
  diagnosis,
  journey_start_begin = "2017-04-01",
  journey_start_end = "2017-06-30"
)
}
\arguments{
\item{journey}{a df containing journeys}

\item{diagnosis}{a df containing diagnosis}

\item{journey_start_begin}{start date for the cohort}

\item{journey_start_end}{end date}
}
\description{
Extracts the journeys with ICD10 heart failure diagnosis
with \code{JOURNEY_START} between \code{journey_start_begin} and \code{journey_start_end}
}
\seealso{
\code{\link{create_hf_snomed_encounters}} for selection of encounter cohort
based on SNOMED codes.

Other heart failure: 
\code{\link{create_hf_icd10_encounters}()},
\code{\link{create_hf_meds_discharge}()},
\code{\link{create_hf_snomed_encounters}()},
\code{\link{extract_echo_procedures}()}
}
\concept{heart failure}
