% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{separation_modes}
\alias{separation_modes}
\title{Definition for separation modes of interest}
\usage{
separation_modes()
}
\description{
Definition for separation modes of interest
}
\seealso{
Other definitions: 
\code{\link{ACS_codes}()},
\code{\link{CCLHD_facility}()},
\code{\link{NSLHD_facility}()},
\code{\link{business_hours}()},
\code{\link{exclusion_criteria_meds}()},
\code{\link{hf_icd_codes}()},
\code{\link{internal_transfers}()}
}
\concept{definitions}
