% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{troponin_encounter_order}
\alias{troponin_encounter_order}
\title{CContain troponin orders based on \code{encounter_order}}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 467528 rows and 11 columns.
}
\usage{
troponin_encounter_order
}
\description{
CContain troponin orders based on \code{encounter_order}
}
\keyword{datasets}
