% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{CCLHD_facility}
\alias{CCLHD_facility}
\title{Returns the definition of CCLHD facilities}
\usage{
CCLHD_facility()
}
\value{
a vector of strings
}
\description{
Returns the definition of CCLHD facilities
}
\seealso{
Other definitions: 
\code{\link{ACS_codes}()},
\code{\link{NSLHD_facility}()},
\code{\link{business_hours}()},
\code{\link{exclusion_criteria_meds}()},
\code{\link{hf_icd_codes}()},
\code{\link{internal_transfers}()},
\code{\link{separation_modes}()}
}
\concept{definitions}
