% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{create_trajectory}
\alias{create_trajectory}
\title{Categorises the trajectories of journeys and whether the journey was a transfer or not
Include a separate column for the presence of a coronary artery bypass procedure}
\usage{
create_trajectory(journeys)
}
\arguments{
\item{journeys}{a df containing encounters and their journey keys}
}
\value{
journeys dataframe with the columns, \code{TRANSFER_STATUS} and \code{TRAJECTORY}
}
\description{
Categorises the trajectories of journeys and whether the journey was a transfer or not
Include a separate column for the presence of a coronary artery bypass procedure
}
