% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{NSLHD_facility}
\alias{NSLHD_facility}
\title{Returns the definition of NSLHD facilities}
\usage{
NSLHD_facility()
}
\value{
a vector of strings
}
\description{
Returns the definition of NSLHD facilities
}
\seealso{
Other definitions: 
\code{\link{ACS_codes}()},
\code{\link{CCLHD_facility}()},
\code{\link{business_hours}()},
\code{\link{exclusion_criteria_meds}()},
\code{\link{hf_icd_codes}()},
\code{\link{internal_transfers}()},
\code{\link{separation_modes}()}
}
\concept{definitions}
